modules: create a separate module for the ffi api
